feat: add basic unittests
