+++ requirements.txt
@@ -1,26 +1,37 @@
-git+https://github.com/suggestionsbot/disnake@bot
-git+https://github.com/suggestionsbot/disnake-ext-components.git
 aiodns==3.0.0
 aiohttp==3.8.1
 aiosignal==1.2.0
 alaric==1.1.3
 async-timeout==4.0.2
 attrs==21.4.0
-Bot-Base==1.5.10
+Bot-Base==1.6.0
 Brotli==1.0.9
+causar==0.1.2
 cchardet==2.1.7
 cffi==1.15.0
 charset-normalizer==2.0.12
+disnake @ git+https://github.com/suggestionsbot/disnake@bot
+disnake-ext-components @ git+https://github.com/suggestionsbot/disnake-ext-components.git@c17a76c507c966e97655a7dd7e1ccc082888bfa8
 dnspython==2.2.1
 frozenlist==1.3.0
+function-cooldowns==1.3.1
 humanize==4.2.0
 idna==3.3
+iniconfig==1.1.1
 motor==3.0.0
+motor-stubs==1.7.1
 multidict==6.0.2
 orjson==3.7.5
+packaging==21.3
+pluggy==1.0.0
+py==1.11.0
 pycares==4.2.1
 pycparser==2.21
 pymongo==4.1.1
-yarl==1.7.2
+pyparsing==3.0.9
+pytest==7.1.3
+pytest-asyncio==0.19.0
 python-dotenv==0.20.0
-function-cooldowns==1.3.1
+tomli==2.0.1
+typing_extensions==4.3.0
+yarl==1.7.2
