+++ suggestions/__init__.py
@@ -4,5 +4,6 @@ from .state import State
 from .colors import Colors
 from .emojis import Emojis
 from .bot import SuggestionsBot
+from .main import create_bot
 
 __all__ = ("SuggestionsBot", "State", "Colors", "Emojis", "ErrorCode", "Stats")
