+++ tests/test_bot.py
@@ -1,0 +1,20 @@
+from typing import cast
+
+from causar import Causar
+
+from suggestions import SuggestionsBot
+
+
+async def test_cogs_loaded(causar: Causar):
+    bot: SuggestionsBot = cast(SuggestionsBot, causar.bot)
+
+    cog_names = [
+        "Internal",
+        "GuildConfigCog",
+        "HelpGuildCog",
+        "SuggestionsCog",
+        "UserConfigCog",
+    ]
+    assert len(bot.cogs) == len(cog_names)
+    for cog_name in cog_names:
+        assert cog_name in bot.cogs
